% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenerated.R, R/condition.R
\name{assertion_message}
\alias{assertion_error}
\alias{assertion_message}
\alias{assertion_warning}
\alias{condition}
\alias{condition_error}
\alias{condition_message}
\alias{condition_warning}
\alias{deprecated_error}
\alias{deprecated_message}
\alias{deprecated_warning}
\alias{dimension_error}
\alias{dimension_message}
\alias{dimension_warning}
\alias{future_error}
\alias{future_message}
\alias{future_warning}
\alias{index_error}
\alias{index_message}
\alias{index_warning}
\alias{io_error}
\alias{io_message}
\alias{io_warning}
\alias{length_error}
\alias{length_message}
\alias{length_warning}
\alias{library_error}
\alias{library_message}
\alias{library_warning}
\alias{lookup_error}
\alias{lookup_message}
\alias{lookup_warning}
\alias{missing_error}
\alias{missing_message}
\alias{missing_warning}
\alias{name_error}
\alias{name_message}
\alias{name_warning}
\alias{runtime_error}
\alias{runtime_message}
\alias{runtime_warning}
\alias{type_error}
\alias{type_message}
\alias{type_warning}
\alias{value_error}
\alias{value_message}
\alias{value_warning}
\title{Generate a custom condition}
\usage{
assertion_message(message, call = sys.call(-1L))

deprecated_message(message, call = sys.call(-1L))

dimension_message(message, call = sys.call(-1L))

future_message(message, call = sys.call(-1L))

index_message(message, call = sys.call(-1L))

io_message(message, call = sys.call(-1L))

length_message(message, call = sys.call(-1L))

library_message(message, call = sys.call(-1L))

lookup_message(message, call = sys.call(-1L))

missing_message(message, call = sys.call(-1L))

name_message(message, call = sys.call(-1L))

runtime_message(message, call = sys.call(-1L))

type_message(message, call = sys.call(-1L))

value_message(message, call = sys.call(-1L))

assertion_warning(message, call = sys.call(-1L))

deprecated_warning(message, call = sys.call(-1L))

dimension_warning(message, call = sys.call(-1L))

future_warning(message, call = sys.call(-1L))

index_warning(message, call = sys.call(-1L))

io_warning(message, call = sys.call(-1L))

length_warning(message, call = sys.call(-1L))

library_warning(message, call = sys.call(-1L))

lookup_warning(message, call = sys.call(-1L))

missing_warning(message, call = sys.call(-1L))

name_warning(message, call = sys.call(-1L))

runtime_warning(message, call = sys.call(-1L))

type_warning(message, call = sys.call(-1L))

value_warning(message, call = sys.call(-1L))

assertion_error(message, call = sys.call(-1L))

deprecated_error(message, call = sys.call(-1L))

dimension_error(message, call = sys.call(-1L))

future_error(message, call = sys.call(-1L))

index_error(message, call = sys.call(-1L))

io_error(message, call = sys.call(-1L))

length_error(message, call = sys.call(-1L))

library_error(message, call = sys.call(-1L))

lookup_error(message, call = sys.call(-1L))

missing_error(message, call = sys.call(-1L))

name_error(message, call = sys.call(-1L))

runtime_error(message, call = sys.call(-1L))

type_error(message, call = sys.call(-1L))

value_error(message, call = sys.call(-1L))

condition(type, class = character(0L), message, call = sys.call(-1L))

condition_error(class, message, call = sys.call(-1L))

condition_warning(class, message, call = sys.call(-1L))

condition_message(class, message, call = sys.call(-1L))
}
\arguments{
\item{message}{[\code{character(1)}]\cr
Information about the condition.}

\item{call}{[\code{call} | \code{NULL}]\cr
Call stack.}

\item{type}{[\code{character(1)}]\cr
Should be one of \dQuote{error}, \dQuote{warning} or \dQuote{message}.}

\item{class}{[\code{character}]\cr
Class for the condition.
The functions \code{condition_error}, \code{condition_warning} and \code{condition_message}
automatically append the respective type with an underscore (see example).}
}
\value{
[\code{condition}].
}
\description{
\code{condition} creates a custom condition.
The functions \code{condition_message}, \code{condition_warning} and \code{condition_error}
are specialized to create conditions of type \dQuote{message}, \dQuote{warning} or \dQuote{error},
respectively. Furthermore, the constructors for some standardized conditions are predefined (see details).
}
\details{
The standardized conditions include:
\describe{
  \item{\dQuote{assertion}:}{Assertion (on user input) failed.}
  \item{\dQuote{deprecated}:}{Feature is deprecated.}
  \item{\dQuote{dimension}:}{Wrong dimension.}
  \item{\dQuote{future}:}{Feature is subject to change in the future.}
  \item{\dQuote{index}:}{Subscript out of range.}
  \item{\dQuote{io}:}{File/directory not found or accessible.}
  \item{\dQuote{length}:}{Wrong length.}
  \item{\dQuote{library}:}{Required package not installed.}
  \item{\dQuote{lookup}:}{Named subelement does not exist.}
  \item{\dQuote{missing}:}{Missing values.}
  \item{\dQuote{name}:}{Failed lookup of a global variable.}
  \item{\dQuote{runtime}:}{Something else which does not fit in any other category went wrong.}
  \item{\dQuote{type}:}{Unexpected type/class.}
  \item{\dQuote{value}:}{Inappropriate value.}
}
}
\examples{
# A simple IO error:
e = condition_error("io", "Failed to load file")
print(e)
class(e)

# To signal the condition, use message/warning/stop.
\dontrun{
message(io_message("Failed to load file"))
warning(io_warning("Failed to load file"))
stop(io_error("Failed to load file"))
}

# These are equivalent:
w1 = condition("warning", "dimension_warning", "foo")
w2 = condition_warning("dimension", "foo")
w3 = dimension_warning("foo")

identical(w1, w2)
identical(w2, w3)
}

