% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{conditions-package}
\alias{conditions}
\alias{conditions-package}
\title{conditions: Standardized Conditions for R}
\description{
Implements specialized conditions, i.e., typed errors,
    warnings and messages. Offers a set of standardized conditions (value error,
    deprecated warning, io message, ...) in the fashion of Python's built-in
    exceptions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mllg/conditions}
  \item Report bugs at \url{https://github.com/mllg/conditions/issues}
}

}
