% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autogenerated.R, R/std_conditions.R
\name{assertion_error}
\alias{assertion_error}
\alias{assertion_warning}
\alias{io_error}
\alias{io_warning}
\alias{length_error}
\alias{length_warning}
\alias{library_error}
\alias{library_warning}
\alias{lookup_error}
\alias{lookup_warning}
\alias{missing_error}
\alias{missing_warning}
\alias{name_error}
\alias{name_warning}
\alias{runtime_error}
\alias{runtime_warning}
\alias{std_conditions}
\alias{type_error}
\alias{type_warning}
\alias{value_error}
\alias{value_warning}
\title{Functions to generate standardized conditions}
\usage{
assertion_error(message, call = sys.call(-1L))

io_error(message, call = sys.call(-1L))

length_error(message, call = sys.call(-1L))

library_error(message, call = sys.call(-1L))

lookup_error(message, call = sys.call(-1L))

missing_error(message, call = sys.call(-1L))

name_error(message, call = sys.call(-1L))

runtime_error(message, call = sys.call(-1L))

type_error(message, call = sys.call(-1L))

value_error(message, call = sys.call(-1L))

assertion_warning(message, call = sys.call(-1L))

io_warning(message, call = sys.call(-1L))

length_warning(message, call = sys.call(-1L))

library_warning(message, call = sys.call(-1L))

lookup_warning(message, call = sys.call(-1L))

missing_warning(message, call = sys.call(-1L))

name_warning(message, call = sys.call(-1L))

runtime_warning(message, call = sys.call(-1L))

type_warning(message, call = sys.call(-1L))

value_warning(message, call = sys.call(-1L))
}
\arguments{
\item{message}{[\code{character(1)}]\cr
Condition message.}

\item{call}{[\code{call} || \code{NULL}]\cr
Call stack.}
}
\value{
[\code{\link[base]{condition}}].
}
\description{
Implementations to generate standardized conditions:
\itemize{
  \item{assertion}{Thrown in assertions.}
  \item{type}{Unexpected type/class.}
  \item{length}{Unexpected length.}
  \item{missing}{Unexpected missing values.}
  \item{lookup}{Named subelement does not exist.}
  \item{index}{Subscript out of range.}
  \item{value}{Inappropriate value.}
  \item{name}{Failed lookup of a global variable.}
  \item{io}{File/directory not found or accessible.}
  \item{library}{Required package not installed.}
  \item{runtime}{Something else which does not fit in any other category went wrong.}
  \item{deprecated}{Some features are deprecated.}
  \item{future}{Something will change semantically in the future.}
}
}

