% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errors.R
\name{std_error}
\alias{assertion_error}
\alias{index_error}
\alias{io_error}
\alias{length_error}
\alias{library_error}
\alias{lookup_error}
\alias{missing_error}
\alias{name_error}
\alias{runtime_error}
\alias{std_error}
\alias{type_error}
\alias{value_error}
\title{Functions to generate standardized errors}
\usage{
assertion_error(message, subclass = character(0L), call = sys.call(-1L))

type_error(message, subclass = character(0L), call = sys.call(-1L))

length_error(message, subclass = character(0L), call = sys.call(-1L))

missing_error(message, subclass = character(0L), call = sys.call(-1L))

io_error(message, subclass = character(0L), call = sys.call(-1L))

lookup_error(message, subclass = character(0L), call = sys.call(-1L))

index_error(message, subclass = character(0L), call = sys.call(-1L))

value_error(message, subclass = character(0L), call = sys.call(-1L))

library_error(message, subclass = character(0L), call = sys.call(-1L))

name_error(message, subclass = character(0L), call = sys.call(-1L))

runtime_error(message, subclass = character(0L), call = sys.call(-1L))
}
\arguments{
\item{message}{[\code{character(1)}]\cr
Error message.}

\item{subclass}{[\code{character}]\cr
Additional classes the condition should inherit from.}

\item{call}{[\code{character(1)}]\cr
Call stack.}
}
\value{
[raised exception].
}
\description{
Implementations to throw errors with the following types:
\itemize{
  \item{assertion_error}{Thrown in assertions.}
  \item{type_error}{Unexpected type/class.}
  \item{length_error}{Unexpected length.}
  \item{missing_error}{Unexpected missing values.}
  \item{lookup_error}{Named subelement does not exist.}
  \item{index_error}{Subscript out of range.}
  \item{value_error}{Inappropriate value.}
  \item{name_error}{Failed lookup of a global variable.}
  \item{io_error}{File/directory not found or accessible.}
  \item{library_error}{Required package not installed.}
  \item{runtime_error}{Something else which does not fit in any other category went wrong.}
}
}

