% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warnings.R
\name{std_warning}
\alias{deprecated_warning}
\alias{future_warning}
\alias{runtime_warning}
\alias{std_warning}
\title{Functions to generate standardized warnings}
\usage{
deprecated_warning(message, subclass = character(0L), call = sys.call(-1L))

future_warning(message, subclass = character(0L), call = sys.call(-1L))

runtime_warning(message, subclass = character(0L), call = sys.call(-1L))
}
\arguments{
\item{message}{[\code{character(1)}]\cr
Warning message.}

\item{subclass}{[\code{character}]\cr
Additional classes the condition should inherit from.}

\item{call}{[\code{character(1)}]\cr
Call stack.}
}
\value{
[warning].
}
\description{
Implementation for the following error types:
\itemize{
  \item{deprecated_warning}{Some features are deprecated.}
  \item{future_warning}{Something will change semantically in the future.}
  \item{runtime_warning}{Something else which does not fit in any other category went wrong.}
}
}

