% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\title{Generate a custom condition}
\usage{
condition(type, message, class = character(0L), call = sys.call(-1L))
}
\arguments{
\item{type}{[\code{character(1)}]\cr
Should be one of \dQuote{error}, \dQuote{warning} or \dQuote{message}.}

\item{message}{[\code{character(1)}]\cr
Information about the condition.}

\item{class}{[\code{character}]\cr
Classes for the condition.}

\item{call}{[\code{character(1)}]\cr
Call stack.}
}
\value{
[\code{condition}] inheriting from dQuote{class} and \dQuote{type}.
}
\description{
Helper function to generate a custom \code{\link[base]{condition}}.
}

