% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{conditions-package}
\alias{conditions}
\alias{conditions-package}
\title{Standardized conditions for R}
\description{
Provides standardized conditions and allows to conveniently create own conditions.
Conditions can be caught and handled via `tryCatch`:
}

